requirements.txt
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
pywavelets==1.4.1
matplotlib==3.7.1
scikit-image==0.21.0
cryptography==41.0.1
pytest==7.4.0
PyQt6==6.4.2